+++ blockstack-sdk/src/main/java/org/blockstack/android/sdk/j2v8/Console.kt
@@ -1,10 +1,12 @@
 package org.blockstack.android.sdk.j2v8
 
 import android.util.Log
+import com.eclipsesource.v8.V8Object
 
 
 interface Console {
     fun error(msg: String)
+    fun error(error: V8Object)
     fun warn(msg: String)
     fun debug(msg: String)
     fun log(msg: String)
@@ -15,6 +17,10 @@ class LogConsole : Console {
 
     override fun error(msg: String) {
         Log.e(TAG, msg)
+    }
+
+    override fun error (error: V8Object) {
+        Log.e(TAG, error.toString())
     }
 
     override fun warn(msg: String) {