+++ CHANGELOG.md
@@ -10,6 +10,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- `BlockstackSession.getFileUrl` added, giving access to the read url for a file.
+- Support for 64bit-apks
+
+### Changed
+- Upgraded to blockstack.js 19.1.0
+- Upgraded android tools and plugins
+- Replace custom scheme in example with verified app links
 
 ## [0.4.4] - 2019-03-08